Test with all available supported interpreters
Reconfigure tox.ini to test against all supported interpreters as
described in setup.cfg. Otherwise stating only "py2" and "py3" will
only test against whatever version of py2 and py3 is available on
the caller's system.

Additionaly remove unnecessary "basepython" declarations. It is
better practice to allow the developer to choose their own basepython
locally. This also causes issues with local development as running
`tox -e py27` may result in missing interpreter despite it being
available so it is better to leave it to tox to figure it out.

Exception for coala and flake8. Coala MUST use python3 so explicitly
set it, while flake8 reacts differently depending on py2 or py3 so
explicitly set it to python3.

Change-Id: Ib83b19e101f2c5d7f9095671322662403f4abed3
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>