Initial commit of the ODL BIST tool
This commits introduces the ODL BIST tool. The idea of BIST is to have
a way to:

- validate deployments
- provide health checks
- make easier for us to diagnose issues

As it stands, this currently provides a test runner that allows the
execution of tests that are of type Diagnostic.

The ideosincracy of this tool is to be able to write
the diagnostic self-tests in the same way a unittest would be written,
and taking advantage of the different assert tools that unittest
provides. Thus, leveraging on a library for which tooling already exist,
it is a proven framework with multiple test runners, and is easy to
extend to account for future requirements (e.g. integrate this to a
REST endpoint for plug into monitoring systems).

This commit is providing the framework to start writing the BIST tests
thats should be placed in `odltools/bist/tests/diagnostic/` and should
comply with the pattern `diag_*.py`. To faciliate the execution of a
test run of the diagnostic tests, this framework introduces the odltools
subcommand `bist` that will collect and run the Diagnostic tests located
in that folder.

Change-Id: I094c3d91a0f84ae9ac59fb8f8dac116f299f496c
Signed-off-by: Michel Peterson <michel@redhat.com>