CLI exit with exit codes
Currently whenever the CLI is invoked it always exits with success,
unless an exception is risen. This does not allow to exit with exit
codes that could be used to signal an execution that was not successful.

This patch adds support of exit codes to the odltools CLI. Whatever the
subcommand default function is returned, that will be used as the exit
code. Underneath it is implemented with sys.exit, so its documentation
applies.

Change-Id: I7ae3324a10776f186dc4a3f3febdc5b74449be73
Signed-off-by: Michel Peterson <michel@redhat.com>