Update patch set 1
Patch Set 1:

Yeah, but they are still figuring as such in the LFN database.
The idea behind this is that once a INFO.yaml is in place, the project can clean them up by just modifying this file and merging the patch.

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>